+++ RiskOfChaos/Utilities/BodySkillPair.cs
@@ -1,0 +1,11 @@
+﻿using RoR2;
+
+namespace RiskOfChaos.Utilities
+{
+    public readonly record struct BodySkillPair(BodyIndex BodyIndex, SkillSlot SkillSlot)
+    {
+        public BodySkillPair(string bodyName, SkillSlot slot) : this(BodyCatalog.FindBodyIndex(bodyName), slot)
+        {
+        }
+    }
+}