+++ CHANGELOG.md
@@ -57,6 +57,9 @@
 * Scrap Random Item:
   * Fixed item notifications not showing for non-host players in multiplayer
 
+* Force Activate Random Skill:
+  * Fixed Beetles freezing in place if their secondary is forced
+
 * Misc
   * Fixed strange scrap counting towards game completion in logbook
   * Fixed some effects sometimes placing character VFX at world origin